move miner_impl
