+++ src/rpc/types/mod.rs
@@ -7,7 +7,6 @@
 
 mod address_impl;
 mod deal_impl;
-mod miner_impl;
 mod sector_impl;
 mod tsk_impl;
 
@@ -17,7 +16,7 @@ mod tests;
 use crate::beacon::BeaconEntry;
 use crate::blocks::TipsetKey;
 use crate::libp2p::Multihash;
-use crate::lotus_json::{lotus_json_with_self, HasLotusJson, LotusJson};
+use crate::lotus_json::{lotus_json_with_self, LotusJson};
 use crate::shim::{
     address::Address,
     clock::ChainEpoch,
@@ -30,17 +29,11 @@ use crate::shim::{
 };
 use cid::Cid;
 use fil_actor_interface::market::AllocationID;
-use fil_actor_interface::miner::{DeadlineInfo, MinerInfo};
-use fil_actor_interface::{
-    market::{DealProposal, DealState},
-    miner::MinerPower,
-    power::Claim,
-};
-use fil_actor_miner_state::v12::{BeneficiaryTerm, PendingBeneficiaryChange};
+use fil_actor_interface::market::{DealProposal, DealState};
+use fil_actor_interface::miner::DeadlineInfo;
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
-use fvm_ipld_encoding::{BytesDe, RawBytes};
+use fvm_ipld_encoding::RawBytes;
 use libipld_core::ipld::Ipld;
-use libp2p::PeerId;
 use num_bigint::BigInt;
 use nunny::Vec as NonEmpty;
 use schemars::JsonSchema;
@@ -149,59 +142,6 @@ pub struct ApiTipsetKey(pub Option<TipsetKey>);
 /// See: <https://github.com/filecoin-project/lotus/issues/11461>.
 #[derive(Debug, Eq, PartialEq, Clone)]
 pub struct AddressOrEmpty(pub Option<Address>);
-
-#[derive(Clone, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "PascalCase")]
-pub struct MinerInfoLotusJson {
-    #[schemars(with = "LotusJson<Address>")]
-    #[serde(with = "crate::lotus_json")]
-    pub owner: Address,
-    #[schemars(with = "LotusJson<Address>")]
-    #[serde(with = "crate::lotus_json")]
-    pub worker: Address,
-    #[schemars(with = "LotusJson<Option<Address>>")]
-    pub new_worker: AddressOrEmpty,
-    #[schemars(with = "LotusJson<Vec<Address>>")]
-    #[serde(with = "crate::lotus_json")]
-    pub control_addresses: Vec<Address>, // Must all be ID addresses.
-    pub worker_change_epoch: ChainEpoch,
-    #[schemars(with = "LotusJson<Option<String>>")]
-    #[serde(with = "crate::lotus_json")]
-    pub peer_id: Option<String>,
-    #[schemars(with = "LotusJson<Vec<Vec<u8>>>")]
-    #[serde(with = "crate::lotus_json")]
-    pub multiaddrs: Vec<Vec<u8>>,
-    #[schemars(with = "String")]
-    pub window_po_st_proof_type: fvm_shared2::sector::RegisteredPoStProof,
-    #[schemars(with = "u64")]
-    pub sector_size: fvm_shared2::sector::SectorSize,
-    pub window_po_st_partition_sectors: u64,
-    pub consensus_fault_elapsed: ChainEpoch,
-    #[schemars(with = "LotusJson<Option<Address>>")]
-    #[serde(with = "crate::lotus_json")]
-    pub pending_owner_address: Option<Address>,
-    #[schemars(with = "LotusJson<Address>")]
-    #[serde(with = "crate::lotus_json")]
-    pub beneficiary: Address,
-    #[schemars(with = "LotusJson<BeneficiaryTerm>")]
-    #[serde(with = "crate::lotus_json")]
-    pub beneficiary_term: BeneficiaryTerm,
-    #[schemars(with = "LotusJson<Option<PendingBeneficiaryChange>>")]
-    #[serde(with = "crate::lotus_json")]
-    pub pending_beneficiary_term: Option<PendingBeneficiaryChange>,
-}
-
-#[derive(Clone, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "PascalCase")]
-pub struct MinerPowerLotusJson {
-    #[schemars(with = "LotusJson<Claim>")]
-    #[serde(with = "crate::lotus_json")]
-    miner_power: Claim,
-    #[schemars(with = "LotusJson<Claim>")]
-    #[serde(with = "crate::lotus_json")]
-    total_power: Claim,
-    has_min_power: bool,
-}
 
 #[derive(Serialize, Deserialize, PartialEq, Clone, JsonSchema)]
 #[serde(rename_all = "PascalCase")]
